api/requirements.txt
mediapipe==0.10.21
numpy==1.26.4
Pillow==11.1.0
psycopg2-binary==2.9.10
